requirements.txt
torch
transformers
datasets
peft
accelerate
numpy
pandas
matplotlib
scikit-learn
python-chess